Remove problematic CSS block | #135894
This block attempts to account for the height of the nav bar, so that when a user links to a heading anchor the heading is not obscured by the nav bar. However, it does not work as expected and is inadvertently making it so links immediately preceding each heading cannot be clicked upon.